requirements.txt
numpy==1.21.6
python-igraph==0.9.6
networkx==2.6.3
pandas==1.3.5
matplotlib==3.5.2
scipy==1.7.3
ConfigModel_MCMC==0.2
graph_tool==2.45